Added reddit scrapping using jraw
To set up, first enter username and password into the fields.
They has been omitted for privacy reasons.

Requests the user to enter a subreddit name, it will return the top 2 (projected to change) submissions in the subreddit based on set filters. It shows the user the number of upvotes the post has, the post title and comments in the post.